Add CloudstreamConfigurationProvider
When using apk - "apk is deprecated; replace with runtimeOnly" so we add a new cloudstream configuration option to avoid this, since we don't want runtimeOnly here. We make ApkConfigurationProvider a wrapper around new CloudstreamConfigurationProvider, and deprecate it. Extensions can now use `val cloudstream by configurations cloudstream("com.lagradost:cloudstream3:pre-release")` instead of `apk` while still maintaining backwards compatibility with `apk`.
