+++ src/main/kotlin/com/lagradost/cloudstream3/gradle/configuration/Configurations.kt
@@ -3,7 +3,10 @@ package com.lagradost.cloudstream3.gradle.configuration
 import org.gradle.api.Project
 
 fun registerConfigurations(project: Project) {
-    val providers = arrayOf(ApkConfigurationProvider())
+    val providers = arrayOf(
+        ApkConfigurationProvider(),
+        CloudstreamConfigurationProvider()
+    )
 
     for (provider in providers) {
         project.configurations.register(provider.name) {
